restructure structures folder
