+++ src/pydrostat/structures/structure.py
